move cairo-rs-py functions to patch file and move call to monkey patch
